Add new skill checking for create-group
