+++ src/main/java/seedu/address/logic/commands/AddSkillCommand.java
@@ -17,7 +17,6 @@ import seedu.address.model.Model;
 import seedu.address.model.coursemate.CourseMate;
 import seedu.address.model.coursemate.QueryableCourseMate;
 import seedu.address.model.coursemate.exceptions.CourseMateNotFoundException;
-import seedu.address.model.group.Group;
 import seedu.address.model.skill.Skill;
 
 /**
@@ -59,7 +58,7 @@ public class AddSkillCommand extends Command {
     public CommandResult execute(Model model) throws CommandException {
         requireNonNull(model);
 
-        Set<Skill> newSkills = getNewSkills(model);
+        Set<Skill> newSkills = model.getNewSkills(addSkillDescriptor.skills);
 
         List<CourseMate> courseMateToEditList;
         try {
@@ -83,59 +82,8 @@ public class AddSkillCommand extends Command {
 
         model.setCourseMate(courseMateToEdit, editedCourseMate);
         model.setRecentlyProcessedCourseMate(editedCourseMate);
-        return new CommandResult(messageNewSkill(newSkills) + MESSAGE_SUCCESS,
+        return new CommandResult(Messages.messageNewSkill(newSkills) + MESSAGE_SUCCESS,
                 false, false, true);
-    }
-
-    /**
-     * Creates a warning message for newly added skills that are not in the database.
-     * @param newSkills - The set of skills that are new to the courseMate list.
-     * @return A String containing the warning message.
-     */
-    public static String messageNewSkill(Set<Skill> newSkills) {
-        if (newSkills.size() == 0) {
-            return "";
-        }
-
-        StringBuilder sb = new StringBuilder();
-        sb.append("WARNING: New skills detected. Please verify for accuracy to avoid any unintended actions: ");
-        int size = newSkills.size();
-        int count = 0;
-        for (Skill skill : newSkills) {
-            sb.append(skill.toString());
-            count++;
-            if (count < size) {
-                sb.append(", ");
-            } else {
-                sb.append("\n");
-            }
-        }
-        return sb.toString();
-    }
-
-    /** Retrieves the newly added skills not in the current courseMate list and group list. */
-    private Set<Skill> getNewSkills(Model model) {
-        List<CourseMate> getCourseMateList = model.getContactList().getCourseMateList();
-        Set<Skill> currentSkills = new HashSet<>();
-        for (CourseMate courseMate : getCourseMateList) {
-            Set<Skill> courseMateSkills = courseMate.getSkills();
-            currentSkills.addAll(courseMateSkills);
-        }
-
-        List<Group> getGroupList = model.getGroupList().getGroupList();
-        for (Group group : getGroupList) {
-            Set<Skill> groupSkills = group.getSkills();
-            currentSkills.addAll(groupSkills);
-        }
-
-        Set<Skill> newSkills = new HashSet<>();
-        Set<Skill> toAddSkill = addSkillDescriptor.getSkills().orElse(new HashSet<>());
-        for (Skill skill : toAddSkill) {
-            if (!currentSkills.contains(skill)) {
-                newSkills.add(skill);
-            }
-        }
-        return newSkills;
     }
 
     /**
